Refactor stock page name
