+++ app/api/stock_routes.py
