+++ react-app/src/components/Stock/index.js
